py_prisma-mu-dns-prod/requirements.txt
prisma-sase>=6.3.1b1
PyYAML>=6.0
python-dotenv>=1.0.0
